requirements.txt
flask
requests
flask-cors
google-generativeai
